RomanEmpireProject/roman_history_stage2/requirements.txt
# requirements.txt
requests>=2.25.1
python-dotenv>=0.19.0
openai>=1.0.0
tiktoken>=0.5.0